Split value into smaller modules; define constrained traits
