+++ contrib/src/constraints/z3.rs
@@ -1,5 +1,5 @@
 use symbolic_stack_machines_core::constraint::*;
-use symbolic_stack_machines_core::value::r#abstract::{AbstractInt, Val};
+use symbolic_stack_machines_core::value::{AbstractInt, Val};
 use z3::ast::{Ast, Bool, Int};
 use z3::{Config, Context, Model, SatResult as Z3SatResult, Solver as Z3InnerSolver};
