+++ contrib/tests/simple_lang.rs
@@ -1,7 +1,7 @@
 use std::rc::Rc;
 use symbolic_stack_machines_core::constraint::Constraint;
 use symbolic_stack_machines_core::memory::{MemOpRecord, MemRecord, ReadOnlyMem};
-use symbolic_stack_machines_core::value::r#abstract::{AbstractInt, Val};
+use symbolic_stack_machines_core::value::{AbstractInt, Val};
 use symbolic_stack_machines_core::{
     instructions::*, machine::*, memory::memory_models::*, stack::*,
 };
